+++ base/src/main/java/io/github/sds100/keymapper/base/keymaps/FingerprintGesturesSupportedUseCase.kt
@@ -1,23 +1,17 @@
 package io.github.sds100.keymapper.base.keymaps
 
-import android.os.Build
 import io.github.sds100.keymapper.data.Keys
 import io.github.sds100.keymapper.data.repositories.PreferenceRepository
 import javax.inject.Inject
 import javax.inject.Singleton
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
 
 @Singleton
 class FingerprintGesturesSupportedUseCaseImpl @Inject constructor(
     private val preferenceRepository: PreferenceRepository,
 ) : FingerprintGesturesSupportedUseCase {
     override val isSupported: Flow<Boolean?> =
-        preferenceRepository.get(Keys.fingerprintGesturesAvailable).map {
-            if (Build.VERSION.SDK_INT < Build.VERSION_CODES.O) return@map false
-
-            it
-        }
+        preferenceRepository.get(Keys.fingerprintGesturesAvailable)
 
     override fun setSupported(supported: Boolean) {
         preferenceRepository.set(Keys.fingerprintGesturesAvailable, supported)